feat(entity describe): allow links to entities
