+++ src/model/mod.rs
@@ -1,2 +1,3 @@
 pub mod entities;
+pub mod fragments;
 pub mod thoughts;
